Fixed the Bidding/Buyout Operation
When one player placed an auction that was too high a sale price for the AH bot to purchase, it would get hung up on this error and not proceed to the next auction from the player. It would also look sequentially at auctions placed so if the bad auction was at the top of the list, it would never move on from that, essentially breaking the AH bot when anyone put up a bad auction. These changes will now make it so that the AH bot alternates which auctions it looks at, skips over the bad ones, and continues to purchase the ones it finds worthy of purchase non-sequentially.